Separate periodic electrostatics boundaries
